#450 move hooks selections to map in klimaorte
